+++ CHANGELOG.md
@@ -7,13 +7,14 @@
   - Updated sortable and nestable empty selectors
   - Moved block padding adjustment into themes
   - Unified sortable and nestable change event parameters
+  - Updated top parameter to support vh values + elements (sticky component)
   - Fixed element swapping between grouped lists
   - Fixed grid match items to wrap onto multiple lines
   - Fixed dropdown hide behavior
   - Fixed dropdown prevent show() on active element
   - Fixed stack margin with dynamic added children
   - Fixed form icon margin
-  
+
 ### 2.21.0 (June 01, 2015)
 
   - IMPORTANT: Reworked nestable component